🚚 refactor: 잘못 배치된 범용 UI/검증을 shared로 이동
[변경 내용]
- 범용 UI 4개를 feature에서 shared/ui로 이동:
  - SectionHeader (project/new/ui/shared -> shared/ui)
  - FilterDropDown (project/list/ui -> shared/ui)
  - ProjectManagementSubTitle (project/management/ui -> shared/ui)
  - Dropdown (challenger/ui/shared -> shared/ui)
- 범용 폼 검증 primitive(emailSchema, codeSchema, idSchema, passwordSchema,
  nicknameSchema)를 signup/validation에서 shared/lib/validationSchemas로 분리.
  signup은 shared에서 import 후 합성 스키마 구성 및 하위 호환 re-export
- 소비자 import 경로(절대/상대) 갱신, auth/useEmailVerification은 emailSchema를
  shared에서 직접 참조

[변경 이유]
- 이 컴포넌트/스키마들은 도메인 의존이 전혀 없는 범용 요소인데 특정 feature에
  있어 다른 feature가 참조하며 cross-feature(features->features) 위반을 유발
- SectionHeader는 application 3곳 + routes 2곳, Dropdown은 challenger + project,
  검증 primitive는 auth/signup/login이 공유 -> 모두 shared가 올바른 위치
- boundaries 위반 27 -> 20건 (-7): application->project 4,
  project->challenger 1, auth->signup 1 등 해소

[고려 사항]
- signup/validation은 범용 primitive와 signup 전용 합성 스키마(signUpSchema,
  oauthSignUpSchema)가 혼재. primitive만 shared로 분리하고 합성은 signup 잔류
- 검증 primitive 추출 시 주석 처리된 idSchema 정의가 함께 딸려와 중복 선언 발생
  -> tsc가 포착, 제거
- storage(혼합 types.ts 의존), useSchools(SchoolNameListResponse 타입 충돌),
  auth/model/types(대형 컨트랙트) 이동은 파일 분리 비용 대비 이득 낮아 보류
- 남은 20건은 대부분 정당한 feature 조합(settings의 계정/인증 관리 -> auth,
  application/matching의 프로젝트 카드 표시 -> project). 페이지 레벨 재조립 또는
  수용이 필요한 별도 영역
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
