+++ src/features/application/ui/ChallengerStatsSection.tsx
@@ -1,6 +1,6 @@
-import { SectionHeader } from "@/features/project/new/ui/shared/SectionHeader"
 import PersonGraphicIcon from "@/shared/assets/icon/people/PersonGraphicIcon"
 import { cn } from "@/shared/lib/utils"
+import { SectionHeader } from "@/shared/ui/SectionHeader"
 
 import { RankBar } from "./RankBar"
 import { RoundDonutChart } from "./RoundDonutChart"
