+++ src/features/project/new/ui/recruit-info/RecruitInfoForm.tsx
@@ -17,6 +17,7 @@ import { Button } from "@/shared/ui/Button"
 import { Counter } from "@/shared/ui/Counter"
 import { OptionButton } from "@/shared/ui/option-button/OptionButton"
 import { OptionButtonGroup } from "@/shared/ui/option-button/OptionButtonGroup"
+import { SectionHeader } from "@/shared/ui/SectionHeader"
 import { useToastStore } from "@/shared/ui/toast/useToastStore"
 
 import {
@@ -25,7 +26,6 @@ import {
   type RoleState,
   useProjectRegisterStore,
 } from "../../model/useProjectRegisterStore"
-import { SectionHeader } from "../shared/SectionHeader"
 
 const ROLES: {
   key: RoleKey
