+++ src/shared/ui/SectionHeader.tsx
